+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# v6.4.0 - 2023-01-06
 
 Added
   * AZ-837: Add Update Management Center